+++ src/ccr/mod.rs
@@ -1,0 +1,5 @@
+mod crecord;
+mod crequest;
+
+pub use crecord::ConfidentialComputeRecord;
+pub use crequest::ConfidentialComputeRequest;
